dashboard/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyyaml==6.0.2
httpx==0.28.1
psutil==6.1.1
websockets==14.1
python-multipart==0.0.19
prometheus_client==0.21.1
